@@ -7,33 +7,30 @@
   "license": "Apache-2.0",
   "scripts": {
     "start": "concurrently --kill-others \"npx y-websocket\" \"ng serve\"",
-    "build": "ng build --configuration production --progress=false --source-map=false",
-    "analyze": "ng build --configuration production --stats-json && webpack-bundle-analyzer dist/stats.json",
+    "build": "ng build --configuration=production --progress=false --source-map=false",
+    "analyze": "ng build --configuration=production --stats-json && webpack-bundle-analyzer dist/stats.json",
     "test": "ng test",
-    "test:ci": "ng test --watch=false --progress=false --browsers=ChromeHeadless",
+    "test:ci": "node --max_old_space_size=4096 ./node_modules/@angular/cli/bin/ng test --watch=false --progress=false --browsers=ChromeHeadless",
     "prettier:fix": "prettier --write ./src",
     "lint": "ng lint",
     "eslint:fix": "yarn eslint --fix ./src",
     "format:fix": "yarn prettier-eslint --write \"src/**/*.{ts,js,html,scss,less,json}\"",
     "format:ci": "yarn prettier-eslint --list-different \"src/**/*.{ts,js,html,scss,less,json}\" && yarn eslint ./src",
-    "postinstall": "node git-version.js && ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points && node ./decorate-angular-cli.js"
+    "postinstall": "node git-version.js && node ./decorate-angular-cli.js"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.9",
-    "@angular/cdk": "~13.3.9",
-    "@angular/common": "~13.3.9",
-    "@angular/compiler": "~13.3.9",
-    "@angular/core": "~13.3.9",
-    "@angular/forms": "~13.3.9",
-    "@angular/material": "~13.3.9",
-    "@angular/platform-browser": "~13.3.9",
-    "@angular/platform-browser-dynamic": "~13.3.9",
-    "@angular/router": "~13.3.9",
+    "@angular/animations": "16.2.12",
+    "@angular/cdk": "16.2.12",
+    "@angular/common": "16.2.12",
+    "@angular/compiler": "16.2.12",
+    "@angular/core": "16.2.12",
+    "@angular/forms": "16.2.12",
+    "@angular/localize": "16.2.12",
+    "@angular/platform-browser": "16.2.12",
+    "@angular/platform-browser-dynamic": "16.2.12",
+    "@angular/router": "16.2.12",
     "@auth0/angular-jwt": "5.1.0",
-    "@deck.gl/core": "~8.9.19",
-    "@deck.gl/layers": "~8.9.19",
-    "@deck.gl/mapbox": "~8.9.19",
     "@loaders.gl/core": "~3.4.2",
     "@luma.gl/core": "~8.5.20",
     "@ngneat/until-destroy": "8.1.4",
@@ -80,50 +71,50 @@
     "tinyqueue": "2.0.3",
     "tslib": "2.3.1",
     "uuid": "8.3.2",
+    "vscode": "npm:@codingame/monaco-vscode-api@~1.76.6",
     "vscode-ws-jsonrpc": "3.1.0",
     "y-monaco": "0.1.5",
     "y-protocols": "1.0.5",
     "y-quill": "0.1.5",
     "y-websocket": "1.4.0",
     "yjs": "13.5.41",
-    "zone.js": "0.11.4"
+    "zone.js": "0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.9",
-    "@angular-eslint/eslint-plugin": "13.1.0",
-    "@angular-eslint/eslint-plugin-template": "13.1.0",
-    "@angular-eslint/template-parser": "13.1.0",
-    "@angular/cli": "~13.3.9",
-    "@angular/compiler-cli": "~13.3.9",
-    "@nrwl/angular": "14.4.6",
-    "@nrwl/cli": "14.4.6",
+    "@angular-builders/custom-webpack": "16.0.1",
+    "@angular-devkit/build-angular": "16.2.12",
+    "@angular-eslint/eslint-plugin": "16.3.1",
+    "@angular-eslint/eslint-plugin-template": "16.3.1",
+    "@angular-eslint/template-parser": "16.3.1",
+    "@angular/cli": "16.2.12",
+    "@angular/compiler-cli": "16.2.12",
+    "@nrwl/cli": "15.9.7",
     "@nrwl/nx-cloud": "16.1.1",
+    "@nx/angular": "18.0.4",
     "@types/backbone": "1.4.15",
-    "@types/c3": "0.7.8",
-    "@types/d3": "6.7.5",
-    "@types/d3-cloud": "1.2.5",
     "@types/dagre": "0.7.47",
     "@types/file-saver": "2.0.5",
     "@types/graphlib": "2.1.8",
     "@types/jasmine": "3.10.3",
     "@types/json-schema": "7.0.9",
     "@types/lodash": "4.14.179",
-    "@types/node": "16.11.7",
+    "@types/node": "~18.15.5",
     "@types/quill": "2.0.9",
     "@types/uuid": "8.3.4",
-    "@typescript-eslint/eslint-plugin": "5.14.0",
-    "@typescript-eslint/parser": "5.14.0",
+    "@typescript-eslint/eslint-plugin": "7.0.2",
+    "@typescript-eslint/parser": "7.0.2",
+    "babel-plugin-dynamic-import-node": "^2.3.3",
     "concurrently": "7.4.0",
-    "eslint": "8.29.0",
-    "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-jsdoc": "46.8.1",
+    "eslint": "8.57.0",
+    "eslint-plugin-import": "2.29.1",
+    "eslint-plugin-jsdoc": "48.2.0",
     "eslint-plugin-prefer-arrow": "1.2.3",
-    "eslint-plugin-prettier": "4.2.1",
-    "eslint-plugin-rxjs": "5.0.2",
-    "eslint-plugin-rxjs-angular": "2.0.0",
+    "eslint-plugin-prettier": "5.1.3",
+    "eslint-plugin-rxjs": "5.0.3",
+    "eslint-plugin-rxjs-angular": "2.0.1",
     "fs-extra": "10.0.1",
     "git-describe": "4.1.0",
-    "jasmine-core": "3.6.0",
+    "jasmine-core": "3.8.0",
     "jasmine-spec-reporter": "5.0.2",
     "karma": "6.3.17",
     "karma-chrome-launcher": "3.1.0",
@@ -131,15 +122,16 @@
     "karma-jasmine": "4.0.1",
     "karma-jasmine-html-reporter": "1.7.0",
     "nodecat": "2.0.0",
-    "nx": "14.4.6",
+    "nx": "16.10.0",
     "nz-tslint-rules": "0.901.2",
-    "prettier": "2.8.8",
-    "prettier-eslint-cli": "7.1.0",
+    "prettier": "3.2.5",
+    "prettier-eslint-cli": "8.0.1",
     "rxjs-marbles": "7.0.1",
-    "sass": "1.49.9",
+    "sass": "1.71.1",
     "ts-node": "4.1.0",
-    "typescript": "4.6.4",
-    "webpack-bundle-analyzer": "4.5.0"
+    "typescript": "5.1.6",
+    "webpack-bundle-analyzer": "4.5.0",
+    "style-loader": "~3.3.3"
   },
   "browserslist": [
     "defaults",
