@@ -1,6 +1,5 @@
 package edu.uci.ics.amber.engine.architecture.controller
 
-import edu.uci.ics.amber.engine.architecture.scheduling.RegionPlan
 import edu.uci.ics.texera.workflow.common.WorkflowContext
 import edu.uci.ics.texera.workflow.common.workflow.{LogicalPlan, PhysicalPlan}
 
@@ -8,6 +7,5 @@ case class Workflow(
     context: WorkflowContext,
     originalLogicalPlan: LogicalPlan,
     logicalPlan: LogicalPlan,
-    physicalPlan: PhysicalPlan,
-    regionPlan: RegionPlan
+    physicalPlan: PhysicalPlan
 ) extends java.io.Serializable {}
