@@ -22,7 +22,7 @@ trait LinkWorkersHandler {
 
   registerHandler[LinkWorkers, Unit] { (msg, sender) =>
     {
-      val region = cp.workflow.regionPlan
+      val region = cp.workflowScheduler.regionPlan
         .getRegionOfLink(msg.link)
       val resourceConfig = region.resourceConfig.get
       val linkConfig = resourceConfig.linkConfigs(msg.link)
