@@ -10,7 +10,7 @@ import edu.uci.ics.amber.engine.architecture.common.WorkflowActor.{
   NetworkMessage
 }
 import edu.uci.ics.amber.engine.architecture.controller.Controller.WorkflowRecoveryStatus
-import edu.uci.ics.amber.engine.architecture.controller.{Controller, ControllerConfig, Workflow}
+import edu.uci.ics.amber.engine.architecture.controller.{Controller, ControllerConfig}
 import edu.uci.ics.amber.engine.common.ambermessage.WorkflowMessage.getInMemSize
 import edu.uci.ics.amber.engine.common.AmberLogging
 import edu.uci.ics.amber.engine.common.ambermessage.{
@@ -28,12 +28,20 @@ import edu.uci.ics.amber.engine.common.client.ClientActor.{
 import edu.uci.ics.amber.engine.common.rpc.AsyncRPCClient.{ControlInvocation, ReturnInvocation}
 import edu.uci.ics.amber.engine.common.rpc.AsyncRPCServer.ControlCommand
 import edu.uci.ics.amber.engine.common.virtualidentity.{ActorVirtualIdentity, ChannelIdentity}
+import edu.uci.ics.texera.workflow.common.WorkflowContext
+import edu.uci.ics.texera.workflow.common.storage.OpResultStorage
+import edu.uci.ics.texera.workflow.common.workflow.PhysicalPlan
 
 import scala.collection.mutable
 
 // TODO: Rename or refactor it since it has mixed duties (send/receive messages, execute callbacks)
 private[client] object ClientActor {
-  case class InitializeRequest(workflow: Workflow, controllerConfig: ControllerConfig)
+  case class InitializeRequest(
+      workflowContext: WorkflowContext,
+      physicalPlan: PhysicalPlan,
+      opResultStorage: OpResultStorage,
+      controllerConfig: ControllerConfig
+  )
   case class ObservableRequest(pf: PartialFunction[Any, Unit])
   case class ClosureRequest[T](closure: () => T)
   case class CommandRequest(command: ControlCommand[_], promise: Promise[Any])
@@ -57,9 +65,11 @@ private[client] class ClientActor extends Actor with AmberLogging {
   }
 
   override def receive: Receive = {
-    case InitializeRequest(workflow, controllerConfig) =>
+    case InitializeRequest(workflowContext, physicalPlan, opResultStorage, controllerConfig) =>
       assert(controller == null)
-      controller = context.actorOf(Controller.props(workflow, controllerConfig))
+      controller = context.actorOf(
+        Controller.props(workflowContext, physicalPlan, opResultStorage, controllerConfig)
+      )
       sender() ! Ack
     case CreditRequest(channelId: ChannelIdentity) =>
       sender() ! CreditResponse(channelId, getQueuedCredit(channelId))
