@@ -1,6 +1,7 @@
 package edu.uci.ics.texera.web.service
 
 import edu.uci.ics.amber.engine.architecture.controller.Workflow
+import edu.uci.ics.amber.engine.architecture.controller.promisehandlers.ReconfigureHandler.Reconfigure
 import edu.uci.ics.amber.engine.architecture.controller.promisehandlers.ModifyLogicHandler.ModifyLogic
 import edu.uci.ics.amber.engine.architecture.worker.promisehandlers.ModifyOperatorLogicHandler.WorkerModifyLogicComplete
 import edu.uci.ics.amber.engine.common.AmberConfig
@@ -99,21 +100,12 @@ class ExecutionReconfigurationService(
         client.sendAsync(ModifyLogic(reconfig._1, reconfig._2))
       })
     } else {
-      val epochMarkers = FriesReconfigurationAlgorithm.scheduleReconfigurations(
-        workflow.physicalPlan,
-        workflow.regionPlan,
-        reconfigurations,
-        reconfigurationId
+      client.sendAsync(Reconfigure(reconfigurations, reconfigurationId))
+
+      // clear all un-scheduled reconfigurations, start a new reconfiguration ID
+      stateStore.reconfigurationStore.updateState(_ =>
+        ExecutionReconfigurationStore(Some(reconfigurationId))
       )
-      epochMarkers.foreach(epoch => {
-        client.sendAsync(epoch)
-      })
     }
-
-    // clear all un-scheduled reconfigurations, start a new reconfiguration ID
-    stateStore.reconfigurationStore.updateState(old =>
-      ExecutionReconfigurationStore(Some(reconfigurationId))
-    )
   }
-
 }
