@@ -1,5 +1,6 @@
 package edu.uci.ics.texera.web.service
 
+import com.google.protobuf.timestamp.Timestamp
 import com.typesafe.scalalogging.LazyLogging
 import edu.uci.ics.amber.engine.architecture.controller.ControllerConfig
 import edu.uci.ics.amber.engine.architecture.worker.WorkflowWorker.{
@@ -15,8 +16,11 @@ import edu.uci.ics.amber.engine.common.virtualidentity.{
 import edu.uci.ics.texera.web.model.websocket.event.TexeraWebSocketEvent
 import edu.uci.ics.texera.web.model.websocket.request.WorkflowExecuteRequest
 import edu.uci.ics.texera.web.service.WorkflowService.mkWorkflowStateId
-import edu.uci.ics.texera.web.storage.WorkflowStateStore
-import edu.uci.ics.texera.web.workflowruntimestate.WorkflowAggregatedState.COMPLETED
+import edu.uci.ics.texera.web.storage.ExecutionStateStore.updateWorkflowState
+import edu.uci.ics.texera.web.storage.{ExecutionStateStore, WorkflowStateStore}
+import edu.uci.ics.texera.web.workflowruntimestate.FatalErrorType.EXECUTION_FAILURE
+import edu.uci.ics.texera.web.workflowruntimestate.WorkflowAggregatedState.{COMPLETED, FAILED}
+import edu.uci.ics.texera.web.workflowruntimestate.WorkflowFatalError
 import edu.uci.ics.texera.web.{SubscriptionManager, WorkflowLifecycleManager}
 import edu.uci.ics.texera.workflow.common.WorkflowContext
 import edu.uci.ics.texera.workflow.common.storage.OpResultStorage
@@ -28,6 +32,7 @@ import play.api.libs.json.Json
 
 import java.util.concurrent.ConcurrentHashMap
 import java.net.URI
+import java.time.Instant
 import scala.jdk.CollectionConverters.IterableHasAsScala
 
 object WorkflowService {
@@ -185,19 +190,44 @@ class WorkflowService(
       }
     }
 
-    val execution = new WorkflowExecutionService(
-      controllerConf,
-      workflowContext,
-      resultService,
-      req,
-      lastCompletedLogicalPlan
-    )
-
-    lifeCycleManager.registerCleanUpOnStateChange(execution.executionStateStore)
-    executionService.onNext(execution)
-    if (execution.executionStateStore.metadataStore.getState.fatalErrors.isEmpty) {
-      execution.startWorkflow()
+    val executionStateStore = new ExecutionStateStore()
+    val errorHandler: Throwable => Unit = { t =>
+      {
+        logger.error("error during execution", t)
+        executionStateStore.statsStore.updateState(stats =>
+          stats.withEndTimeStamp(System.currentTimeMillis())
+        )
+        executionStateStore.metadataStore.updateState { metadataStore =>
+          updateWorkflowState(FAILED, metadataStore).addFatalErrors(
+            WorkflowFatalError(
+              EXECUTION_FAILURE,
+              Timestamp(Instant.now),
+              t.toString,
+              t.getStackTrace.mkString("\n"),
+              "unknown operator"
+            )
+          )
+        }
+      }
     }
+
+    try {
+      val execution = new WorkflowExecutionService(
+        controllerConf,
+        workflowContext,
+        resultService,
+        req,
+        executionStateStore,
+        errorHandler,
+        lastCompletedLogicalPlan
+      )
+      lifeCycleManager.registerCleanUpOnStateChange(executionStateStore)
+      executionService.onNext(execution)
+      execution.startWorkflow()
+    } catch {
+      case e: Throwable => errorHandler(e)
+    }
+
   }
 
   def convertToJson(frontendVersion: String): String = {
