@@ -1,9 +1,11 @@
 package edu.uci.ics.amber.engine.architecture.scheduling
 
+import edu.uci.ics.amber.engine.architecture.controller.WorkflowScheduler
 import edu.uci.ics.amber.engine.common.virtualidentity.OperatorIdentity
 import edu.uci.ics.amber.engine.common.workflow.PortIdentity
 import edu.uci.ics.amber.engine.e2e.TestOperators
 import edu.uci.ics.amber.engine.e2e.TestUtils.buildWorkflow
+import edu.uci.ics.texera.workflow.common.storage.OpResultStorage
 import edu.uci.ics.texera.workflow.common.workflow.LogicalLink
 import edu.uci.ics.texera.workflow.operators.split.SplitOpDesc
 import edu.uci.ics.texera.workflow.operators.udf.python.{
@@ -19,6 +21,7 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
     val headerlessCsvOpDesc = TestOperators.headerlessSmallCsvScanOpDesc()
     val keywordOpDesc = TestOperators.keywordSearchOpDesc("column-1", "Asia")
     val sink = TestOperators.sinkOpDesc()
+    val resultStorage = new OpResultStorage()
     val workflow = buildWorkflow(
       List(headerlessCsvOpDesc, keywordOpDesc, sink),
       List(
@@ -34,23 +37,26 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
           sink.operatorIdentifier,
           PortIdentity(0)
         )
-      )
+      ),
+      resultStorage
     )
+    val workflowScheduler = new WorkflowScheduler(workflow.context, resultStorage)
+    workflowScheduler.updateSchedule(workflow.physicalPlan)
 
-    assert(workflow.regionPlan.regions.size == 1)
-    workflow.regionPlan.topologicalIterator().zip(Iterator(3)).foreach {
+    assert(workflowScheduler.regionPlan.regions.size == 1)
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(3)).foreach {
       case (regionId, opCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getOperators.size == opCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getOperators.size == opCount)
     }
 
-    workflow.regionPlan.topologicalIterator().zip(Iterator(2)).foreach {
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(2)).foreach {
       case (regionId, linkCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getLinks.size == linkCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getLinks.size == linkCount)
     }
 
-    workflow.regionPlan.topologicalIterator().zip(Iterator(4)).foreach {
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(4)).foreach {
       case (regionId, portCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getPorts.size == portCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getPorts.size == portCount)
     }
   }
 
@@ -59,6 +65,7 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
     val headerlessCsvOpDesc2 = TestOperators.headerlessSmallCsvScanOpDesc()
     val joinOpDesc = TestOperators.joinOpDesc("column-1", "column-1")
     val sink = TestOperators.sinkOpDesc()
+    val resultStorage = new OpResultStorage()
     val workflow = buildWorkflow(
       List(
         headerlessCsvOpDesc1,
@@ -85,26 +92,30 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
           sink.operatorIdentifier,
           PortIdentity()
         )
-      )
+      ),
+      resultStorage
     )
 
-    assert(workflow.regionPlan.regions.size == 2)
-    workflow.regionPlan.topologicalIterator().zip(Iterator(3, 3)).foreach {
+    val workflowScheduler = new WorkflowScheduler(workflow.context, resultStorage)
+    workflowScheduler.updateSchedule(workflow.physicalPlan)
+
+    assert(workflowScheduler.regionPlan.regions.size == 2)
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(3, 3)).foreach {
       case (regionId, opCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getOperators.size == opCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getOperators.size == opCount)
     }
 
-    workflow.regionPlan.topologicalIterator().zip(Iterator(2, 2)).foreach {
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(2, 2)).foreach {
       case (regionId, linkCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getLinks.size == linkCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getLinks.size == linkCount)
     }
 
-    workflow.regionPlan.topologicalIterator().zip(Iterator(4, 4)).foreach {
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(4, 4)).foreach {
       case (regionId, portCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getPorts.size == portCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getPorts.size == portCount)
     }
 
-    val buildRegion = workflow.regionPlan.regions
+    val buildRegion = workflowScheduler.regionPlan.regions
       .find(region =>
         region.getOperators
           .map(_.id)
@@ -113,7 +124,7 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
           )
       )
       .get
-    val probeRegion = workflow.regionPlan.regions
+    val probeRegion = workflowScheduler.regionPlan.regions
       .find(region =>
         region.getOperators
           .map(_.id)
@@ -123,7 +134,9 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
       )
       .get
 
-    assert(workflow.regionPlan.regionLinks.contains(RegionLink(buildRegion.id, probeRegion.id)))
+    assert(
+      workflowScheduler.regionPlan.regionLinks.contains(RegionLink(buildRegion.id, probeRegion.id))
+    )
 
   }
 
@@ -132,6 +145,7 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
     val keywordOpDesc = TestOperators.keywordSearchOpDesc("column-1", "Asia")
     val joinOpDesc = TestOperators.joinOpDesc("column-1", "column-1")
     val sink = TestOperators.sinkOpDesc()
+    val resultStorage = new OpResultStorage()
     val workflow = buildWorkflow(
       List(
         headerlessCsvOpDesc1,
@@ -164,22 +178,27 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
           sink.operatorIdentifier,
           PortIdentity()
         )
-      )
+      ),
+      resultStorage
     )
-    assert(workflow.regionPlan.regions.size == 2)
-    workflow.regionPlan.topologicalIterator().zip(Iterator(5, 3)).foreach {
+
+    val workflowScheduler = new WorkflowScheduler(workflow.context, resultStorage)
+    workflowScheduler.updateSchedule(workflow.physicalPlan)
+
+    assert(workflowScheduler.regionPlan.regions.size == 2)
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(5, 3)).foreach {
       case (regionId, opCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getOperators.size == opCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getOperators.size == opCount)
     }
 
-    workflow.regionPlan.topologicalIterator().zip(Iterator(4, 2)).foreach {
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(4, 2)).foreach {
       case (regionId, linkCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getLinks.size == linkCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getLinks.size == linkCount)
     }
 
-    workflow.regionPlan.topologicalIterator().zip(Iterator(7, 4)).foreach {
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(7, 4)).foreach {
       case (regionId, portCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getPorts.size == portCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getPorts.size == portCount)
     }
   }
 
@@ -189,6 +208,7 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
     val hashJoin1 = TestOperators.joinOpDesc("column-1", "Region")
     val hashJoin2 = TestOperators.joinOpDesc("column-2", "Country")
     val sink = TestOperators.sinkOpDesc()
+    val resultStorage = new OpResultStorage()
     val workflow = buildWorkflow(
       List(
         buildCsv,
@@ -228,22 +248,26 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
           sink.operatorIdentifier,
           PortIdentity()
         )
-      )
+      ),
+      resultStorage
     )
-    assert(workflow.regionPlan.regions.size == 2)
-    workflow.regionPlan.topologicalIterator().zip(Iterator(5, 4)).foreach {
+
+    val workflowScheduler = new WorkflowScheduler(workflow.context, resultStorage)
+    workflowScheduler.updateSchedule(workflow.physicalPlan)
+    assert(workflowScheduler.regionPlan.regions.size == 2)
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(5, 4)).foreach {
       case (regionId, opCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getOperators.size == opCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getOperators.size == opCount)
     }
 
-    workflow.regionPlan.topologicalIterator().zip(Iterator(4, 3)).foreach {
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(4, 3)).foreach {
       case (regionId, linkCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getLinks.size == linkCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getLinks.size == linkCount)
     }
 
-    workflow.regionPlan.topologicalIterator().zip(Iterator(7, 6)).foreach {
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(7, 6)).foreach {
       case (regionId, portCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getPorts.size == portCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getPorts.size == portCount)
     }
   }
 
@@ -253,6 +277,7 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
     val training = new PythonUDFOpDescV2()
     val inference = new DualInputPortsPythonUDFOpDescV2()
     val sink = TestOperators.sinkOpDesc()
+    val resultStorage = new OpResultStorage()
     val workflow = buildWorkflow(
       List(
         csv,
@@ -292,22 +317,27 @@ class ExpansionGreedyRegionPlanGeneratorSpec extends AnyFlatSpec with MockFactor
           sink.operatorIdentifier,
           PortIdentity()
         )
-      )
+      ),
+      resultStorage
     )
-    assert(workflow.regionPlan.regions.size == 2)
-    workflow.regionPlan.topologicalIterator().zip(Iterator(5, 3)).foreach {
+
+    val workflowScheduler = new WorkflowScheduler(workflow.context, resultStorage)
+    workflowScheduler.updateSchedule(workflow.physicalPlan)
+
+    assert(workflowScheduler.regionPlan.regions.size == 2)
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(5, 3)).foreach {
       case (regionId, opCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getOperators.size == opCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getOperators.size == opCount)
     }
 
-    workflow.regionPlan.topologicalIterator().zip(Iterator(4, 2)).foreach {
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(4, 2)).foreach {
       case (regionId, linkCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getLinks.size == linkCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getLinks.size == linkCount)
     }
 
-    workflow.regionPlan.topologicalIterator().zip(Iterator(8, 4)).foreach {
+    workflowScheduler.regionPlan.topologicalIterator().zip(Iterator(8, 4)).foreach {
       case (regionId, portCount) =>
-        assert(workflow.regionPlan.getRegion(regionId).getPorts.size == portCount)
+        assert(workflowScheduler.regionPlan.getRegion(regionId).getPorts.size == portCount)
     }
   }
 
