@@ -1,6 +1,6 @@
 package edu.uci.ics.amber.engine.e2e
 
-import edu.uci.ics.amber.engine.architecture.controller.{ControllerConfig, Workflow}
+import edu.uci.ics.amber.engine.architecture.controller.Workflow
 import edu.uci.ics.texera.web.model.websocket.request.LogicalPlanPojo
 import edu.uci.ics.texera.web.storage.ExecutionStateStore
 import edu.uci.ics.texera.workflow.common.WorkflowContext
@@ -13,7 +13,7 @@ object TestUtils {
   def buildWorkflow(
       operators: List[LogicalOp],
       links: List[LogicalLink],
-      resultStorage: OpResultStorage = new OpResultStorage()
+      resultStorage: OpResultStorage
   ): Workflow = {
     val context = new WorkflowContext()
     val workflowCompiler = new WorkflowCompiler(
@@ -23,8 +23,7 @@ object TestUtils {
       LogicalPlanPojo(operators, links, List(), List()),
       resultStorage,
       None,
-      new ExecutionStateStore(),
-      ControllerConfig.default
+      new ExecutionStateStore()
     )
   }
 
