@@ -35,11 +35,8 @@ export class HomeComponent implements OnInit {
         }),
         untilDestroyed(this)
       )
-      .subscribe(
-        // The new page will append to the bottom of the page, and the original page does not remove, zone solves this issue
-        Zone.current.wrap(() => {
-          this.router.navigateByUrl(this.route.snapshot.queryParams["returnUrl"] || "/dashboard/workflow");
-        }, "")
+      .subscribe(() =>
+        this.router.navigateByUrl(this.route.snapshot.queryParams["returnUrl"] || "/dashboard/workflow")
       );
   }
 }
